+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## UNRELEASED
+- Horse with Leather Horse Armor can now walk on Powder Snow like a player with Leather Boots.
 - Backported Leash knot untying and breaking sounds from a recent minecraft version.
 - Dismounting from a horse will now place a player towards their view direction rather than on left/right side of a horse.
 - Fixed horse sometimes running back when the player dismounts it.